Removed ZNC support.
ZNC support is no longer something I would consider to be a feature of a
general use IRC bot. It's a feature specific to a very small niche of
people who run bouncers with so many users they NEED a bot to help
manage it.

It's also one more thing that has to work across multiple versions of
third party software, something I don't have the time or will to keep up
with.

ZNC support is still be available in the 'znc' git tag, but this tag
will receive no maintenance or updates.